Fix includes for WIN32
Saldo_freegame_new: Mean = 559.41M, stddev = 87.39M

+++ src/Tafel.cpp
@@ -402,7 +402,7 @@ void CTafelData::Randomize(SLONG Day) {
         // Players might set a min value
         citiesToPick = std::max(citiesToPick, Sim.Options.OptionRentOfficeMinAvailable);
         // Also, we don't want to set more than the availableCities or more than 7 (because UI doesn't manage it yet).
-        citiesToPick = std::min({citiesToPick, (ULONG)availableCities.size(), 7U});
+        citiesToPick = std::min({citiesToPick, (ULONG)availableCities.size(), (ULONG)7U});
 
         AT_Log("citiesToPick = %d (numTries = %d), availableCities = %d", citiesToPick, numTries, availableCities.size());
